+++ specification/ml/evaluate_data_frame/MlEvaluateDataFrameRequest.ts
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-import { DataFrameEvaluationContainer } from '@ml/_types/DataFrameEvaluation'
+import { DataframeEvaluationContainer } from '@ml/_types/DataframeEvaluation'
 import { RequestBase } from '@_types/Base'
 import { IndexName } from '@_types/common'
 import { QueryContainer } from '@_types/query_dsl/abstractions'
@@ -33,7 +33,7 @@ export interface Request extends RequestBase {
      * Defines the type of evaluation you want to perform.
      * @doc_url https://www.elastic.co/guide/en/elasticsearch/reference/current/evaluate-dfanalytics.html#ml-evaluate-dfanalytics-resources
      */
-    evaluation: DataFrameEvaluationContainer
+    evaluation: DataframeEvaluationContainer
     /**
      * Defines the index in which the evaluation will be performed.
      */
