+++ specification/ml/explain_data_frame_analytics/MlExplainDataFrameAnalyticsResponse.ts
@@ -20,7 +20,7 @@
 import {
   DataFrameAnalyticsFieldSelection,
   DataFrameAnalyticsMemoryEstimation
-} from '@ml/_types/DataFrameAnalytics'
+} from '@ml/_types/DataframeAnalytics'
 
 export class Response {
   body: {
