+++ specification/ml/update_data_frame_analytics/MlUpdateDataFrameAnalyticsResponse.ts
@@ -22,7 +22,7 @@ import {
   DataFrameAnalysisContainer,
   DataFrameAnalyticsDestination,
   DataFrameAnalyticsSource
-} from '@ml/_types/DataFrameAnalytics'
+} from '@ml/_types/DataframeAnalytics'
 import { ByteSize, Id, VersionString } from '@_types/common'
 import { integer, long } from '@_types/Numeric'
